BAU: add local startup option to startup.sh

+++ startup.sh
@@ -2,10 +2,14 @@
 set -eu
 
 CLEAN=0
-while getopts "c" opt; do
+LOCAL=0
+while getopts "cl" opt; do
   case ${opt} in
     c)
         CLEAN=1
+      ;;
+    l)
+        LOCAL=1
       ;;
     *)
         usage
@@ -20,12 +24,16 @@ if [ $CLEAN == "1" ]; then
   rm -rf node_modules
 fi
 
-running_count=$(docker ps --filter status=running | grep "di-auth-frontend-dev" | wc -l | awk '{ print $1 }')
-if [ ${running_count} -ne 0 ]; then
-  echo "Stopping di-auth-frontend-dev..."
-  docker stop di-auth-frontend-dev
-  docker rm di-auth-frontend-dev --force
+if [ $LOCAL == "1" ]; then
+  echo "Starting di-auth-frontend-dev on local..."
+  yarn install && yarn run copy-build && yarn run dev
+else
+  running_count=$(docker ps --filter status=running | grep "di-auth-frontend-dev" | wc -l | awk '{ print $1 }')
+  if [ ${running_count} -ne 0 ]; then
+    echo "Stopping di-auth-frontend-dev..."
+    docker stop di-auth-frontend-dev
+    docker rm di-auth-frontend-dev --force
+  fi
+  echo "Starting di-auth-frontend-dev in docker..."
+  docker run --name di-auth-frontend-dev -dp 3000:3000 -w /app -v "$(pwd):/app"  node:15-alpine sh -c "yarn install && yarn run copy-build && yarn run dev"
 fi
-
-echo "Starting di-auth-frontend-dev in frontend..."
-docker run --name di-auth-frontend-dev -dp 3000:3000 -w /app -v "$(pwd):/app"  node:15-alpine sh -c "yarn install && yarn run copy-build && yarn run dev"
